storage: add stub endpoint to get the recovery key for TPM/FDE
This introduces a new endpoint meant to be used by the desktop installer at the
end of an installation in the TPM/FDE flow when requesting a recovery key.

* GET /v2/storage/core_boot_recovery_key

For now, the returned key is not usable, but in the future we will get it from
snapd instead - which will associate it to the encrypted device.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>