storage: reflect the step name in system action response snapd type
Previously, only the SystemActionRequest that had "setup-storage-encryption" as
the step had an interesting response, so calling the response type
"SystemActionResponse" was fine. However, we will now need more than one type
for the response, in particular for the step "generate-recovery-key".

Let's have the name of the step in the type name.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>