make: fix coverage report generation
pytest-cov, which uses coverage.py underneath, had trouble writing a
report in the TICS default location .coverage/cobertura.xml due to
coverage.py expecting to use .coverage as a temporary file in sqlite3
format.  If we set an alternate location for the data_file instead of
.coverage, we can generate the report.

Also, pytest-cov is cool with .coverage not existing as a directory so
simplify the makefile a bit.