storage: generate the recovery later, so it doesn't expire
When calling /v2/systems/{label} with action="generate-recovery-key", a
recovery key is created with an expire duration of 5 minutes (defined in
snapd).

Snapd expects that the /v2/systems/{label} with action="finish-install" gets
called before the key expires.

In case the installer doesn't, we end up with:

    File "subiquity/server/snapd/api.py", line 142, in post_and_wait
      raise aiohttp.ClientError(result.err)
  aiohttp.client_exceptions.ClientError: cannot perform the following tasks:
  - Finish setup of run system for "enhanced-secureboot-desktop" (recovery key has expired)

We now move the "generate-recovery-key" step just before the "finish-install"
step. This should drastically lower the window for key expiration.

The drawback is that subiquity's /v2/storage/core_boot_recovery_key endpoint
can now only be used late during the installation, but UDB consumes it at the
very end so that is ok.

LP: #2114738

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>