storage: ensure variations info is populated, when marking storage configured
When the storage model is marked configured, it expects the variation info to
be populated. This is typically true when installing Ubuntu. However, with
automated tests, storage.configured() sometimes gets called while we are still
in the middle of _examine_systems.

If that happens, the variation info is not populated and we end up with:

    File "subiquity/server/controllers/filesystem.py", line 373, in configured
      self.set_info_for_capability(GuidedCapability.DIRECT)
    File "subiquity/server/controllers/filesystem.py", line 779, in set_info_for_capability
      raise Exception("could not find variation for {}".format(capability))
  Exception: could not find variation for GuidedCapability.DIRECT

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>