snapd: set Content-Type header appropriately when POSTing to snapd
When POSTing to Snapd, we send it JSON data. Previously, we did not specify a
Content-Type header, which is fine with snapd. We now rely on requests to set
the Content-Type header appropriately (i.e., to application/json).

When debugging, it makes it easier to work with captured data. It also makes it
easier to implement a mock version of snapd with flask.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>