Revert "oem: fix kernel specification"
This reverts commit 081dc12f90c2802c5cb3c5462c199b35ef54e5da.

In the above commit, we tried to fix the name of the OEM kernel package stored
in the kernel model. We figured that "linux-oem-somerville-treckoo-meta-24.04"
is not a real package name and replaced it with "oem-somerville-treckoo-meta".

The latter does exist, however, this is still wrong because it isn't a kernel
metapackage, it is an OEM metapackage.

LP: #2114780
Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>