oem: fix kernel metapackage name for OEM kernel
On certified devices, when we are offline and the OEM metapackage declares that
it should be used with the OEM kernel (i.e., Ubuntu-Oem-Kernel-Flavour: oem),
we store the name of the kernel package to install in the kernel model.

However, the flavor_to_pkgname() function was used incorrectly. We were meant
to give it the name of a flavor (e.g., oem, generic, ...) but instead we gave
it the name of the OEM metapackage.

As a consequence, we ended up trying to install kernel package with names such
as "linux-oem-somerville-treckoo-meta-24.04". Instead it should have been
"linux-oem-24.04".

LP: #2114780

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 0c997768bc4a0a94ef123298d37fe953d6499976)