storage: fix inability to edit a partition having a GPT name
Subiquity does not support changing the name of a GPT partition. However,
Subiquity intends to support preserving the name if it is already set. To do
so, the user should set the "name" field of the /storage/v2/edit_partition
request to the original name.

Unfortunately, when doing so, Subiquity still fails with the following error:

  ValueError: edit_partition does not support changing partition name

Indeed, there was a flow when comparing the original name and the new name.

The Partition API object has a "name" field which corresponds to the (GPT)
partition name. But in the Partition model object, the "name" field is
different. It is inherited from curtin and controls the presence of a
/dev/disk/by-dname/<name> symlink.

LP: #2117025

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit bdb5ba96720a7ab685473dca8b3f586c79f64305)