storage: fix wrong test to check if the name field is specified
In theory, when calling /storage/v2/edit_partition with no "name" field, we
request to preserve the name of the partition (if any).

However, the test to check if the "name" field is set was wrong. This lead to
the following error when the "name" field is not present but the partition
already has a name set:

  ValueError: edit_partition does not support changing partition name

Fixed by changing the test.

LP: #2117025

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 2e3979667535bb5f30f539a6e3fc957fb49da0f4)