network: fix missing with_wlan_listener var when running console-conf
When we introduced the --no-wlan-listener option, we relied on its value
in subiquitycore.controllers.network.BaseNetworkController as part of
the probe_network method.

However, BaseNetworkController is not just used by Subiquity server. It
is also used by Subiquity client and console-conf, where the
--no-wlan-listener option does not exist.

The subiquity client does not invoke probe_network so that's okayish.
But console-conf does and crashes with:

    File "subiquitycore/core.py", line 119, in start_controllers
      controller.start()
      ~~~~~~~~~~~~~~~~^^
    File "subiquitycore/controllers/network.py", line 184, in start
      with_wlan_listener=self.app.opts.with_wlan_listener,
                         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  AttributeError: 'Namespace' object has no attribute 'with_wlan_listener'

Fixed by:
 * having a with_wlan_listener variable at the BaseNetworkController
   level, so that calling probe_network is always fine.
 * overriding the value in the Subiquity server's network controller.

Generally speaking, it seems wrong to use app.opts anywhere in
subiquitycore.

LP: #2122027

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>