kvm-test: add USB passthrough support
This allows for example to give Wi-Fi support to the guest if the host
has a spare USB Wi-Fi adapter.

  $ kvm-test --usb-passthrough $bus $addr

Example:

  $ lsusb
  [...]
  Bus 001 Device 013: ID 0bda:8179 Realtek Semiconductor Corp. RTL8188EUS 802.11n Wireless Network Adapter
  [...]

  $ kvm-test --usb-passthrough 1 13

Note that the current user must have write access to the device in
question.

You can give it access temporarily using:

  $ sudo chown [...] /dev/bus/usb/$bus/$addr

e.g.:

  $ sudo chown olivier /dev/bus/usb/001/013

or more persistently using a udev rule such as:

 SUBSYSTEM=="usb", ATTRS{idVendor}=="...", ATTRS{idProduct}=="...", OWNER="..."

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>