network: fix network loss if Wi-Fi IP configured before AP settings
From a netplan standpoint, it is invalid to declare a Wi-Fi
configuration without defining the access points settings.

And if a user of Subquity decides to configure the Wi-Fi IP settings
before the "Wi-Fi" (i.e., AP) settings, the netplan configuration
produced becomes invalid. As soon as the user validates the IP settings,
the network is lost immediately (for all interfaces).

The logs show that "netplan apply" fails with:

 > Error in network definition: {iface}: No access points defined

Furthermore, if using console-conf instead of Subiquity, the exception
is uncaught and terminates the console-conf process.

Fixed by not producing the netplan configuration when a Wi-Fi device is
incompletely configured.

LP: #2121854

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>