doc: fix debconf-selections example from OpenSSH
The example we had for debconf-selection using
openssh-server/permit-root-login was wrong. We were assuming that setting the
value to false meant "PermitRootLogin no". But it actually means
"PermitRootLogin yes" which is super confusing.

In hindsight, using openssh as an example for debconf-selections was a bad idea
but dropping it completely from our documentation would be unfair to users who
have been copying our examples. Let's keep it around a while longer but make it
correct.

LP: #2128863

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>