install: move identity user creation before first boot
So far, user creation was done on first boot using cloud-init no matter what.

While this is a valid option, it makes it harder for people to write
postinstall scripts.

It also makes it hard for the installer to create ZFS user datasets since the
first boot has no user created.

When users are handled by the identity model, We now create the user in the
postinstall phase. This makes it possible for admins to use the newly created
user in the postinstall scripts.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit bca943a597e10f28acc0f929ca7aaeff180fed80)

LP: #2090834