runner: prevent environment expansion when using systemd-run
Unbeknownst to me, systemd-run attempts to do environment variable expansion in
a way similar to the shell. Therefore, executing the following:

 * $ systemd-run echo '$HOME'
 * $ systemd-run echo 'a${foo}'

does not result in '$HOME' or 'a${foo}' being printed. Instead the "HOME" and
"foo" variables are expanded if they exist (otherwise, they are expanded to an
empty string).

Since systemd 254, we can pass --expand-environment=no to systemd-run to
disable environment variable expansion. Unfortunately, this would break the
command runner on jammy.

The other backward compatible way is to escape the dollar signs, by replacing
single dollar occurrences by double dollars.

 * $ systemd-run echo '$$HOME'    => prints '$HOME'
 * $ systemd-run echo 'a$${foo}'  => prints 'a${foo}'

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>