identity: add autoinstall support for setting groups
In the identity section, one can now specify the groups directive.

By default, the user is added to:
 * the sudo group (hardcoded in Subiquity)
 * the admin group (hardcoded in Subiquity)
 * a group named after the user (automatically created by useradd(8))

The groups directive can be used to either:

 * get the user added to a list of supplementary groups,
 * get the user added to a list of groups, ignoring groups that are
   hardcoded in Subiquity.

To ignore hardcoded groups, use:

  # Two equivalent syntaxes for specifying the list of groups the user should
  # belong to:
  groups: [sudo, admin, lpadmin]
  groups: {override: [sudo, admin, lpadmin]}

And for supplementary groups:

  # Add the user to the wireshark group in addition to the default
  groups: {append: [wireshark]}

Note that in any case, the user will be added to a group named after them.
This group is automatically created by useradd.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>