snapdapi: add mechanism to disambiguate the return type
Many snapd endpoints return different types of data depending on the
arguments passed. This is hard to work with with our current client API
implementation in Subiquity.

This patch adds support for declaring return types of endpoints as unions of
different types, and narrow down which type we expect when calling the
endpoint.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>