shutdown: use async command runner for shutdown
We used to lean a run_command to execute the call to shutdown/reboot. This is
fine but using the asynchronous alternative feels more natural now.

There should be no noticeable change since calling `systemctl reboot` is an
asynchronous operation anyway (i.e., it returns as soon as the reboot or
poweroff operation is initiated).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>