runner: accept to pipe in input, provided we also capture the output
We can now pass input when calling runner.run() using the "input" argument.
That said, because of how reading stdin is implemented in systemd-run, using
input="something" will cause stdout and stderr to not end up in the journal.

Therefore, we currently only allow input="something" in combination with
capture=True.

In the future, it would be good to support one without the other.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>