identity: prevent encrypted password from being logged by systemd
In general, if we pass an encrypted password using the --password option of
useradd, then we need to make sure that the invocation is not logged anywhere.

This is challenging to do, especially in Subiquity where we run useradd through
systemd-run. Indeed the command automatically gets recorded in the journal as a
transient unit.

To avoid the problem, we now call useradd without the password option, and then
immediately call chpasswd to update the password, using stdin.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>